mv preview to nodes_api
